requirements.txt
langchain==0.0.284
python-dotenv==1.0.0
streamlit==1.22.0
python-magic==0.4.27
python-magic-bin==0.4.14
huggingface-hub==0.17.1
